pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cogbot"
version = "0.1.0"
description = "Human-like cognitive AI agent framework."
readme = "readme.md"
requires-python = ">=3.9"
dependencies = [
  "cryptography",
  "groq",
  "numpy",
  "openai",
  "python-dotenv",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["cogbot*"]
